chore: refactor carbon scoped token provider tests to RTL
