+++ src/style/design-tokens/carbon-scoped-tokens-provider/carbon-scoped-tokens-provider.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { mount } from "enzyme";
+import { render } from "@testing-library/react";
 import { ThemeProvider } from "styled-components";
 import { noTheme, sageTheme } from "../../themes";
 import CarbonScopedTokensProvider from "./carbon-scoped-tokens-provider.component";
@@ -8,7 +8,7 @@ jest.mock("../../../__internal__/utils/helpers/guid", () => () => "guid");
 
 describe("CarbonScopedTokensProvider", () => {
   it("should render css variables for all themes", () => {
-    mount(
+    render(
       <>
         <ThemeProvider theme={noTheme}>
           <CarbonScopedTokensProvider />
@@ -19,6 +19,10 @@ describe("CarbonScopedTokensProvider", () => {
       </>
     );
 
-    expect(document.styleSheets[0].cssRules).toMatchSnapshot();
+    // Collect all the CSS stylesheets content for snapshot testing
+    const cssRules = Array.from(document.styleSheets).flatMap((sheet) =>
+      Array.from(sheet.cssRules).map((rule) => rule.cssText)
+    );
+    expect(cssRules).toMatchSnapshot();
   });
 });
